Move common dictionary service to node-core
