+++ packages/types/src/project.ts
@@ -175,6 +175,7 @@ export interface SubstrateDatasourceProcessor<
 export interface DictionaryQueryCondition {
   field: string;
   value: string;
+  matcher?: string; // defaults to "equalTo", use "contains" for JSON
 }
 
 export interface DictionaryQueryEntry {
